+++ walk-in-interview/pom.xml
@@ -67,6 +67,7 @@
                 <artifactId>appengine-maven-plugin</artifactId>
                 <version>2.2.0</version>
                 <configuration>
+                    <port>3000</port>
                     <deploy.projectId>google.com:walk-in-interview</deploy.projectId>
                     <deploy.version>1</deploy.version>
                 </configuration>
